requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.19.0
langchain==0.1.13
langchain-experimental==0.0.55
langchain-google-genai==1.0.1
python-dotenv==1.0.1
tabulate==0.9.0
streamlit-float==0.4.0